ハンバーガーメニューの遷移時の作動
